feat: add reply callback step controller
